\name{ALT}
\alias{ALT}
\docType{data}
\title{ ALT Integer Constant}
\description{
   References # of columns comparing estimators to each other
}
\usage{ALT}
\format{
  The format is:
 num 21
}
\details{
     Just a constant that we don't even use anymore
}
\source{
    Alan Lenarcic
}
\references{
    TwoLasso For Bayes Adapted Selection, Lenarcic JCGS 2009
}
\examples{
DeclareGlobals();
ALT
## maybe str(ALT) ; plot(ALT) ...
}
\keyword{datasets}
