\name{DeclareGlobals}
\alias{DeclareGlobals}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ DeclareGlobals }
\description{
 Declares Some Global Values
}
\usage{
 DeclareGlobals();
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  
}
\details{
   None Needed
}
\value{
}
\references{ "TwoLasso for Bayes Adapted Selection", Lenarcic,
  Journal of Computational and Graphical Statistics, Submitted 2009. }
\author{ Alan Lenarcic }
\note{ Loads Global Values
}
\seealso{ SecondLoadfunction }
\examples{
DeclareGlobals();
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ methods }
\keyword{ arith }% __ONLY ONE__ keyword per line
\keyword{ internal }
