\name{EstimatorColNames}
\alias{EstimatorColNames}
\docType{data}
\title{Column Names for Data Table}
\description{
  Column Names for Printing LAtex Tables
}
\usage{EstimatorColNames}
\format{
   Character Vector
}
\details{
  None Needed
}
\source{
  Alan Lenarcic's Harvard Thesis 2009
}
\references{
}
\examples{
##data(EstimatorColNames).
}
\keyword{datasets}
\keyword{internal}
