\name{FailToFitEMOb}
\alias{FailToFitEMOb}
\docType{data}
\title{ An Internal List of Objects for which EM did not successfuly fit}
\description{
  Error Checking Failure List
}
\usage{FailToFitEMOb}
\format{
 list()
}
\details{
  List grows as EMRIDGE, EMLARS, XLOperation, etc. return rejection lists.
}
\source{
  Alan Lenarcic's 2009 Harvard Thesis
}
\seealso{ OnFails, FailToFitSim }
\references{
}
\examples{
##data(FailToFitEMOb)
}
\keyword{datasets}
\keyword{ error }
\keyword{ debugging }
