\name{FailToFitSim}
\alias{FailToFitSim}
\docType{data}
\title{ Simulation Data sets that could not be fit}
\description{
  Error Checking List of failed datasets.
}
\usage{FailToFitSim}
\format{
  The format is:
 list()
}
\details{
    Contains the data sets that EMLARS, EMRIDGE, XLOperation failed to fit.
}
\source{
  Alan Lenarcic's Harvard 2009 Thesis
}
\seealso{ OnFails, FailToFitEMOb }
\references{
   
}
\examples{
##data(FailToFitSim)
## maybe str(FailToFitSim) ; plot(FailToFitSim) ...
}
\keyword{datasets}
\keyword{error}
\keyword{debugging}
