\name{FunctionPlot}
\alias{FunctionPlot}
\docType{data}
\title{A character string for function plotting}
\description{
  Just Some Character Strings
}
\usage{FunctionPlot}
\format{
  The format is Type II errors on the x-axis and Type I errors on the yaxis.
}
\details{
      Not really more
}
\source{
  None
}
\references{
  None
}
\examples{
##data(FunctionPlot)
## maybe str(FunctionPlot) ; plot(FunctionPlot) ...
}
\keyword{datasets}
