\name{GiveGiveMe}
\alias{GiveGiveMe}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ GiveGiveMe }
\description{
  GiveGiveMe Comparison
}
\usage{
GiveGiveMe(Arrs)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Arrs}{ List of Variables to compare amongst to each other }
}
\details{
  Internal Method tests simulated estimates against each other for 
    active coherence
}
\value{

}
\references{ Lenarcic Thesis 2009}
\author{Alan Lenarcic }
\note{ None
}
\seealso{ GetForMeAAA2, GetForMeAll2 }
\examples{


}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ methods }
\keyword{ internal }% __ONLY ONE__ keyword per line
