\name{InitiateS}
\alias{InitiateS}
\docType{data}
\title{ Initiate String }
\description{
  String going with NomenClatureFunction
}
\usage{InitiateS}
\format{
  The format is:
 chr "BFOUR"
}
\details{
}
\seealso{NamingClenture, NamingClenture2}
\source{
}
\references{
}
\examples{
##data(InitiateS)
## maybe str(InitiateS) ; plot(InitiateS) ...
}
\keyword{internal}
