\name{InitiateS2}
\alias{InitiateS2}
\docType{data}
\title{ InitiateS2 String for Datasets}
\description{
  A String going into NomenClature functions
}
\usage{InitiateS2}
\format{
  The format is:
 chr "BSIXB8"
}
\details{
}
\source{
}
\references{
}
\examples{
##data(InitiateS2)
## maybe str(InitiateS2) ; plot(InitiateS2) ...
}
\keyword{datasets}
\keyword{internal}
