\name{LoadSavedOutPutDirectory}
\alias{LoadSavedOutPutDirectory}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Load SavedOutput Directory }
\description{
  Creates a Directory for saving database simulations.
}
\usage{
LoadSavedOutPutDirectory()
}
%- maybe also 'usage' for other objects documented here.
\details{
}
\value{
   A String, and sets up other files to copy directory
}
\references{ }
\author{ Alan Lenarcic }
\note{ 
}
\seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }
\keyword{ documentation }% __ONLY ONE__ keyword per line
