\name{LoadSavedTableDirectory}
\alias{LoadSavedTableDirectory}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Place to Save Latex Tables }
\description{
   Finds Directory to Save Latex Output Tables
}
\usage{
LoadSavedTableDirectory()
}
%- maybe also 'usage' for other objects documented here.
\details{
}
\value{
   A String, and sets up other files to copy directory
}
\references{ }
\author{ Alan Lenarcic }
\note{ 
}
\seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }
\keyword{ documentation }% __ONLY ONE__ keyword per line
