\name{LoadTwoCC}
\alias{LoadTwoCC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Load TwoLars Directoy }
\description{
  Loads CC TwoLasso.dll into directory.
}
\usage{
LoadTwoCC()
}
%- maybe also 'usage' for other objects documented here.
\details{
  Internally Loads CC code
}
\value{

}
\references{ Lenarcic 2009 Thesis }
\author{ Alan Lenarcic }
\note{ Not Much to say, get filenames riht.
}

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
LoadTwoCC()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }
\keyword{ IO }% __ONLY ONE__ keyword per line
