\name{meanC}
\alias{meanC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ mean of Columns }
\description{
 Declares Some Global Values
}
\usage{
 meanC(MatX)
}
%- maybe also 'usage' for other objects documented here.
\arguments{

}
\details{
   None Needed
}
\value{
}
\references{ "TwoLasso for Bayes Adapted Selection", Lenarcic,
  Journal of Computational and Graphical Statistics, Submitted 2009. }
\author{ Alan Lenarcic }
\note{ Same as colMeans
}
\seealso{ colMeans }
\examples{
matX <- matrix(1:12, 3,4);
meanC(matX);
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ methods }
\keyword{ arith }% __ONLY ONE__ keyword per line
\keyword{ internal }
