\name{n1}
\alias{n1}
\docType{data}
\title{ An integer referring to default sample sizes for simulations}
\description{
  An integer
}
\usage{n1}
\format{
  The format is:
 num 100
}
\details{
   Default Sample size for simulations
}
\source{
  Alan Lenarcic
}
\references{
"TwoLasso for Bayes Adapted Selection", Lenarcic,
  Journal of Computational and Graphical Statistics, Submitted 2009.
}
\examples{
DeclareGlobals();
print(n1);
## maybe str(n1) ; plot(n1) ...
}
\keyword{datasets}
