\name{Pnames}
\alias{Pnames}
\docType{data}
\title{ Internal Names of 7 estimators}
\description{
  An old list of estimators
}
\usage{Pnames}
\format{
  The format is:
 chr [1:7] 
}
\details{
  List of Estimators, no longer used.
}
\source{
  Alan Lenarcic
}
\references{
  Alan Lenarcic's 2009 Harvard Thesis Chapter 2

   "TwoLasso for Bayes Adapted Selection", Lenarcic, Journal of
   Computational and Graphical Statistics, Submitted 2009.
}
\examples{
Pnames
## maybe str(Pnames) ; plot(Pnames) ...
}
\keyword{datasets}
\keyword{internal}
