\name{PnamesV2}
\alias{PnamesV2}
\docType{data}
\title{ PnamesV2 vector of all estimators }
\description{
  Length 11 list of Estimator String names
}
\usage{PnamesV2}
\format{
  The format is:
 chr
}
\details{
   None Really
}
\source{Alan Lenarcic
}
\references{ Alan Lenarcic's 2009 Harvard Thesis Chapter 2

   "TwoLasso for Bayes Adapted Selection", Lenarcic, Journal of
   Computational and Graphical Statistics, Submitted 2009.
}
\examples{
PnamesV2
## maybe str(PnamesV2) ; plot(PnamesV2) ...
}
\keyword{datasets}
\keyword{ internal }
