\name{SecondLoadfunction}
\alias{SecondLoadfunction}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Second Function To Load }
\description{
 Sets up Error Token Lists
}
\usage{
SecondLoadfunction()
}
%- maybe also 'usage' for other objects documented here.
\arguments{

}
\details{
   None Needed
}
\value{
}
\references{ "TwoLasso for Bayes Adapted Selection", Lenarcic,
  Journal of Computational and Graphical Statistics, Submitted 2009. }
\author{ Alan Lenarcic }
\note{ Same as colMeans
}
\seealso{ DeclareGlobals }
\examples{
SecondLoadfunction();
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ arith }% __ONLY ONE__ keyword per line
\keyword{ internal }
