\name{TopPlot}
\alias{TopPlot}
\docType{data}
\title{ TopPlot}
\description{
  A character array of words to go on top of some plots
}
\usage{TopPlot}

\details{
       Not So Interesting
}
\source{
  
}
\references{
  Alan Lenarcic Thesis 2009
}
\examples{
data(TopPlot)
## maybe str(TopPlot) ; plot(TopPlot) ...
}
\keyword{datasets}
