\name{TwoDigits}
\alias{TwoDigits}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Identify first two base 10 digits }
\description{
   Identifies first two digits of number
}
\usage{
TwoDigits(Nums, NDigits)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Nums}{ Number to Convert }
  \item{NDigits}{ Number of Digits to Allow }
}
\details{
   None Necessary
}
\value{
    Returns first Two significant digits of the number
}
\references{ TwoLasso for Bayes Adapted Selection, Lenarcic JCGS 2009 }
\author{ Alan Lenarcic }
\note{  None Necessary
}
\seealso{ rd0 }
\examples{
TwoDigits(349,2);
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }
\keyword{ IO }% __ONLY ONE__ keyword per line
