\name{WhatGoesEachBoxBB}
\alias{WhatGoesEachBoxBB}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ What goes in each Box BB }
\description{
  Helper Function for printing to Latex Tables
}
\usage{
    WhatGoesEachBoxBB(tt, MeanTII, SdTII, MeanTI, SdTI, MeanBetaSq, SdBetaSq,
            PercentPer, CompTU, SdCompTU, CompTC, SdCompTC, rndit)
}
\details{
   Called by other functions for "DoAllTheSaving"
   Formats numeric values in Latex tables
}
\value{
  Don't input
}
\references{ Alan Lenarcic's Thesis Harvard 2009 }
\author{ Alan Lenarcic }
\note{ Not Really
}
\seealso{ DoAllTheSaving }
\examples{ 
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }
\keyword{ print }% __ONLY ONE__ keyword per line
